+++ test/ElementInternals.test.js
@@ -8,6 +8,7 @@ import {
 import "../dist/index.js";
 
 let callCount = 0;
+let internalsAvailableInFormAssociatedCallback = false;
 
 class CustomElement extends HTMLElement {
   static get formAssociated() {
@@ -18,9 +19,14 @@ class CustomElement extends HTMLElement {
     super();
     const root = this.attachShadow({ mode: "open" });
     this.internals = this.attachInternals();
-    root.innerHTML = "html";
+    root.innerHTML = "<input>";
 
     this._value = "";
+  }
+
+  connectedCallback() {
+    this.tabIndex = -1;
+    this.input = this.shadowRoot.querySelector("input");
   }
 
   set disabled(disabled) {
@@ -62,6 +68,10 @@ class CustomElement extends HTMLElement {
     return this._value;
   }
 
+  formAssociatedCallback() {
+    internalsAvailableInFormAssociatedCallback = !!this.internals;
+  }
+
   formDisabledCallback() {
     callCount += 1;
   }
@@ -97,8 +107,8 @@ describe("The ElementInternals polyfill", () => {
     });
   });
 
-  describe("inside a custom element without a form", () => {
-    let el, internals;
+  describe("Non-formAssociated elements", () => {
+    let element, internals;
 
     class NotFormAssociated extends HTMLElement {
       constructor() {
@@ -106,41 +116,40 @@ describe("The ElementInternals polyfill", () => {
         this.internals = this.attachInternals();
       }
     }
+
     customElements.define("not-associated", NotFormAssociated);
 
     beforeEach(async () => {
-      el = await fixture(html`<test-el></test-el>`);
-      internals = el.internals;
+      element = await fixture(html`<not-associated></not-associated>`);
     });
 
-    afterEach(async () => fixtureCleanup(el));
+    afterEach(async () => await fixtureCleanup(element));
 
     it("will attach an object to internals even if not form associated", async () => {
-      const nonAssociated = await fixture(
-        html` <not-associated></not-associated>`
-      );
-      expect(nonAssociated.internals).to.exist;
+      expect(element.internals).to.exist;
     });
 
-    it("will return undefined from setFormValue", async () => {
-      expect(internals.setFormValue("foo")).to.equal(undefined);
+    it("will throw from setFormValue", async () => {
+      expect(() => element.internals.setFormValue("foo")).to.throw();
     });
   });
 
   describe("inside a custom element with a form", () => {
-    let form, el, label, button, internals;
+    let form, el, noname, label, button, internals;
 
     beforeEach(async () => {
       form = await fixture(html`
         <form id="form" onsubmit="return false;">
           <label for="foo">Label text</label>
           <test-el name="foo" id="foo"></test-el>
+          <test-el id="noname"></test-el>
           <button type="submit">Submit</button>
         </form>
       `);
       callCount = 0;
       label = form.querySelector("label");
-      el = form.querySelector("test-el");
+      el = form.querySelector("test-el[id=foo]");
+      noname = form.querySelector("test-el[id=noname]");
       button = form.querySelector("button");
       internals = el.internals;
     });
@@ -238,6 +247,17 @@ describe("The ElementInternals polyfill", () => {
       }).to.throw();
     });
 
+    it("will accept ValidityState from a native form input", () => {
+      el.input.required = true;
+      el.input.reportValidity();
+      internals.setValidity(
+        el.input.validity,
+        el.input.validationMessage,
+        el.input
+      );
+      expect(internals.validity.valueMissing).to.be.true;
+    });
+
     it("will return true for willValidate if the field can participate in the form", () => {
       expect(internals.willValidate).to.be.true;
     });
@@ -272,7 +292,7 @@ describe("The ElementInternals polyfill", () => {
       // Lifecycle methods are stripped off at definition time
       // and added elsewhere so we can't use a spy. Instead
       // we're going to look for a side-effect
-      expect(callCount).to.equal(1);
+      expect(callCount).to.equal(2);
     });
 
     it("will cancel form submission if invalid", (done) => {
@@ -305,6 +325,133 @@ describe("The ElementInternals polyfill", () => {
         "This field is required"
       );
       el.internals.checkValidity();
+    });
+
+    it("will call formAssociatedCallback after internals have been set", () => {
+      expect(internalsAvailableInFormAssociatedCallback).to.be.true;
+    });
+
+    it("will not include null values set via setFormValue", () => {
+      internals.setFormValue("test");
+      internals.setFormValue(null);
+      const output = new FormData(form);
+      expect(Array.from(output.keys()).length).to.equal(0);
+    });
+
+    it("will not include undefined values set via setFormValue", () => {
+      internals.setFormValue("test");
+      internals.setFormValue(undefined);
+      const output = new FormData(form);
+      expect(Array.from(output.keys()).length).to.equal(0);
+    });
+
+    it("will include multiple form values passed via FormData to setFormValue", () => {
+      let input;
+      let output;
+      input = new FormData();
+      input.set("first", "1");
+      input.set("second", "2");
+      input.append("second", "22"); // Multi-value keys should also work
+      internals.setFormValue(input);
+      output = new FormData(form);
+      expect(Array.from(output.values()).length).to.equal(3);
+      expect(output.get("first")).to.equal("1");
+      expect(output.getAll("second").length).to.equal(2);
+      input = new FormData();
+      input.set("override", "3");
+      internals.setFormValue(input);
+      output = new FormData(form);
+      expect(Array.from(output.keys()).length).to.equal(1);
+      expect(output.get("override")).to.equal("3");
+    });
+
+    it("will not include form values from elements without a name", () => {
+      noname.internals.setFormValue("noop");
+      const output = new FormData(form);
+      expect(Array.from(output.keys()).length).to.equal(0);
+    });
+
+    it("will include form values from elements without a name if set with FormData", () => {
+      const formData = new FormData();
+      formData.set("formdata", "works");
+      noname.internals.setFormValue(formData);
+      const output = new FormData(form);
+      expect(Array.from(output.keys()).length).to.equal(1);
+      expect(output.get("formdata")).to.equal("works");
+    });
+
+    it("saves a reference to all shadow roots", () => {
+      expect(internals.shadowRoot).to.equal(el.shadowRoot);
+    });
+
+    it("will focus the element if validated with anchor", async () => {
+      internals.setValidity(
+        {
+          customError: true,
+        },
+        "Error message",
+        el.input
+      );
+      internals.reportValidity();
+      expect(document.activeElement).to.equal(el);
+    });
+
+    it("will focus anchor elements in document order on form submission failure", () => {
+      internals.setValidity(
+        {
+          customError: true,
+        },
+        "Error message",
+        el.input
+      );
+      noname.internals.setValidity(
+        {
+          customError: true,
+        },
+        "Error message",
+        noname.input
+      );
+      button.click();
+      expect(document.activeElement).to.equal(el);
+    });
+
+    it("will accept non strings", async () => {
+      internals.setFormValue(["a", "b"]);
+      expect(new FormData(internals.form).get("foo")).to.equal("a,b");
+    });
+
+    it("will accept empty strings", () => {
+      internals.setFormValue("");
+      expect(new FormData(internals.form).get("foo")).to.equal("");
+    });
+  });
+
+  describe("closed shadow root element", () => {
+    let shadowRoot;
+    let el;
+    let internals;
+
+    class ClosedRoot extends HTMLElement {
+      constructor() {
+        super();
+        shadowRoot = this.attachShadow({ mode: "closed" });
+        this.internals = this.attachInternals();
+      }
+    }
+
+    customElements.define("closed-root", ClosedRoot);
+
+    beforeEach(async () => {
+      el = await fixture(html`<closed-root></closed-root>`);
+      internals = el.internals;
+    });
+
+    afterEach(async () => {
+      await fixtureCleanup(el);
+    });
+
+    it("maintains a reference to closed shadow roots", () => {
+      expect(internals.shadowRoot).to.equal(shadowRoot);
     });
   });
 });